Complete exercise
Completed twice, first time my solution was not as elegant
as the solution file, so I reworked it to better understand the
solution. I did not think to text-align the body element, or
realize that text-wrap defaults to wrap; as a result my solution was
more complex than it needed.